refactor: 클래스 네이밍에 맞지 않는 DTO 분리 (#325)
