+++ src/main/java/org/sopt/app/application/auth/dto/RecommendFriendRequest.java
@@ -1,0 +1,14 @@
+package org.sopt.app.application.auth.dto;
+
+import java.util.List;
+import org.sopt.app.application.auth.dto.PlaygroundProfileInfo.RecommendFriendFilter;
+
+public record RecommendFriendRequest(
+        List<Integer> generations,
+        List<RecommendFriendFilter> filters
+) {
+
+    public static RecommendFriendRequest createRecommendFriendRequestByGeneration(List<Integer> generations) {
+        return new RecommendFriendRequest(generations, List.of());
+    }
+}
